Further refactor actions
