+++ src/store/actions/comments.js
@@ -1,0 +1,8 @@
+// import axios from 'axios'
+// import {BASE_URL} from './config.js'
+
+export default {
+  commentsFetch (postId) {
+
+  }
+}
